requirements.txt
pytest-playwright 
requests
faker